+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Duplicate fields in report/finding field editor
 * Allow creating custom notifications in django admin
+* Fix error when formatting risk score in AI agent
 
 
 ## v2025.104 - 2025-12-10